[feature #165476908] add API of admin to get all loan applications, renamed some files for clarifications
