+++ server/app.js
@@ -1,7 +1,7 @@
 /* eslint-disable no-console */
 import express from 'express';
 import bodyParser from 'body-parser';
-import routes from './routes/users';
+import routes from './routes/router';
 
 const app = express();
 
@@ -10,6 +10,7 @@ app.use(bodyParser.json());
 
 // app.use(users);
 app.use('/api/v1/users/', routes);
+app.use('/api/v1/admin/', routes);
 
 const port = process.env.PORT || 4000;
 app.listen(port, () => { console.log(`app is running on ${port}...`); });
